refactor(MapPage): separate MapDisplay component, prepare for separate separate route data layer
