+++ assets/src/components/mapPage/routePropertiesCard.tsx
