+++ assets/tests/components/mapPage.test.tsx
@@ -8,7 +8,6 @@ import useSearchResults from "../../src/hooks/useSearchResults"
 import * as dateTime from "../../src/util/dateTime"
 
 import userEvent from "@testing-library/user-event"
-import { useTripShape } from "../../src/hooks/useShapes"
 import useVehicleForId from "../../src/hooks/useVehicleForId"
 import { useStations } from "../../src/hooks/useStations"
 import { LocationType } from "../../src/models/stopData"
@@ -27,7 +26,6 @@ import {
   searchQueryRunFactory,
   searchQueryVehicleFactory,
 } from "../factories/searchQuery"
-import shapeFactory from "../factories/shape"
 import stopFactory from "../factories/stop"
 import vehicleFactory, {
   randomLocationVehicle,
@@ -40,16 +38,19 @@ import useVehiclesForRoute from "../../src/hooks/useVehiclesForRoute"
 import routeFactory from "../factories/route"
 import { RealDispatchWrapper } from "../testHelpers/wrappers"
 import { VehicleId, VehicleOrGhost } from "../../src/realtime"
-import { RouteId, Shape, TripId } from "../../src/schedule"
+import { RouteId } from "../../src/schedule"
+import useRoutePatterns from "../../src/hooks/useRoutePatterns"
+import { routePatternFactory } from "../factories/routePattern"
+import shape from "../factories/shape"
 
 jest.mock("../../src/hooks/useSearchResults", () => ({
   __esModule: true,
   default: jest.fn(() => null),
 }))
 
-jest.mock("../../src/hooks/useShapes", () => ({
+jest.mock("../../src/hooks/useRoutePatterns", () => ({
   __esModule: true,
-  useTripShape: jest.fn(() => null),
+  default: jest.fn(() => null),
 }))
 
 jest.mock("../../src/hooks/useNearestIntersection", () => ({
@@ -94,9 +95,27 @@ function mockUseVehicleForId(vehicles: VehicleOrGhost[]) {
   mockUseVehicleForIdMap(vehicleIdToVehicleMap)
 }
 
-function mockUseTripShape(map: { [tripId: TripId]: Shape }) {
-  ;(useTripShape as jest.Mock).mockImplementation((tripId) =>
-    [map[tripId!]].filter(Boolean)
+function mockRoutePatternsForVehicles(
+  vehicles: VehicleOrGhost[],
+  params?: { stopCount: number }
+) {
+  const patternIds = Array.from(
+    new Set(
+      vehicles
+        .filter((v) => v.routePatternId != null)
+        .map((v) => v.routePatternId)
+    )
+  )
+
+  ;(useRoutePatterns as jest.Mock).mockReturnValue(
+    patternIds.map((routePatternId) =>
+      routePatternFactory.build({
+        id: routePatternId!,
+        shape: shape.build({
+          stops: stopFactory.buildList(params?.stopCount || 2),
+        }),
+      })
+    )
   )
 }
 
@@ -234,11 +253,7 @@ describe("<MapPage />", () => {
       { runId } = nextVehicle
     ;(useSearchResults as jest.Mock).mockReturnValue(vehicles)
 
-    const [a, b] = shapeFactory.buildList(2)
-    mockUseTripShape({
-      [vehicle.tripId!]: a,
-      [nextVehicle.tripId!]: b,
-    })
+    mockRoutePatternsForVehicles([vehicle, nextVehicle])
 
     mockUseVehicleForId([vehicle, nextVehicle])
     mockUseVehiclesForRouteMap({ [route.id]: vehicles })
@@ -276,7 +291,7 @@ describe("<MapPage />", () => {
     const runId = "clickMe"
     const vehicle = vehicleFactory.build({ runId })
     ;(useSearchResults as jest.Mock).mockReturnValue([vehicle])
-    mockUseTripShape({ [vehicle.tripId!]: shapeFactory.build() })
+    mockRoutePatternsForVehicles([vehicle])
     mockUseVehicleForId([vehicle])
 
     const { container } = render(
@@ -311,8 +326,7 @@ describe("<MapPage />", () => {
       savedQueries: [],
     }
 
-    const shape = shapeFactory.build()
-    mockUseTripShape({ [vehicle.tripId!]: shape })
+    mockRoutePatternsForVehicles([vehicle])
 
     const mockDispatch = jest.fn()
     const { container } = render(
@@ -362,7 +376,7 @@ describe("<MapPage />", () => {
     ;(useSearchResults as jest.Mock).mockReturnValue(vehicles)
     mockUseVehicleForId(vehicles)
     mockUseVehiclesForRouteMap({ [vehicle.routeId!]: vehicles })
-    mockUseTripShape({ [vehicle.tripId!]: shapeFactory.build() })
+    mockRoutePatternsForVehicles([vehicle])
 
     const { container } = render(
       <RealDispatchWrapper
@@ -406,7 +420,6 @@ describe("<MapPage />", () => {
     jest.spyOn(global, "scrollTo").mockImplementationOnce(jest.fn())
     const runId = runIdFactory.build()
     const vehicle = vehicleFactory.build({ runId })
-    const shape = shapeFactory.build()
     const activeSearch: SearchPageState = searchPageStateFactory.build({
       query: searchQueryRunFactory.searchFor(vehicle.runId!).build(),
       selectedEntity: {
@@ -416,7 +429,7 @@ describe("<MapPage />", () => {
     })
 
     ;(useSearchResults as jest.Mock).mockReturnValue([vehicle])
-    mockUseTripShape({ [vehicle.tripId!]: shape })
+    mockRoutePatternsForVehicles([vehicle])
     mockUseVehicleForId([vehicle])
     mockUseVehiclesForRouteMap({
       [vehicle.routeId!]: [vehicle],
@@ -630,13 +643,12 @@ describe("<MapPage />", () => {
         const { id: runId } = RunFactory.build()
         const vehicles = vehicleFactory.buildList(3, { runId }),
           [vehicle] = vehicles
-        const shape = shapeFactory.build()
         const searchPageState = activeSearchPageStateFactory.build({
           query: searchQueryRunFactory.searchFor(vehicle.runId!).build(),
         })
 
         ;(useSearchResults as jest.Mock).mockReturnValue(vehicles)
-        mockUseTripShape({ [vehicle.tripId!]: shape })
+        mockRoutePatternsForVehicles([vehicle])
 
         render(
           <StateDispatchProvider
@@ -664,10 +676,7 @@ describe("<MapPage />", () => {
 
         ;(useSearchResults as jest.Mock).mockReturnValue(vehicles)
 
-        const shape = shapeFactory.build({
-          stops: stopFactory.buildList(3),
-        })
-        mockUseTripShape({ [vehicle.tripId!]: shape })
+        mockRoutePatternsForVehicles([vehicle])
         mockUseVehicleForId([vehicle])
         mockUseVehiclesForRouteMap({
           [vehicle.routeId!]: vehicles,
@@ -755,8 +764,7 @@ describe("<MapPage />", () => {
             [route.id]: selectedRouteVehicles,
           })
 
-          const shape = shapeFactory.build()
-          mockUseTripShape({ [selectedVehicle.tripId!]: shape })
+          mockRoutePatternsForVehicles([selectedVehicle])
 
           render(
             <StateDispatchProvider
@@ -785,14 +793,12 @@ describe("<MapPage />", () => {
             })
 
             const [selectedVehicle] = selectedRouteVehicles
-            const shape = shapeFactory.build({
-              stops: stopFactory.buildList(8),
-            })
+            const stopCount = 8
 
             setHtmlWidthHeightForLeafletMap()
             mockUseVehicleForId([selectedVehicle])
             mockUseVehiclesForRouteMap({ [route.id]: selectedRouteVehicles })
-            mockUseTripShape({ [selectedVehicle.tripId!]: shape })
+            mockRoutePatternsForVehicles([selectedVehicle], { stopCount: 8 })
 
             const { container } = render(
               <StateDispatchProvider
@@ -809,7 +815,7 @@ describe("<MapPage />", () => {
 
             expect(
               container.querySelectorAll(".m-vehicle-map__stop")
-            ).toHaveLength(shape.stops?.length || 0)
+            ).toHaveLength(stopCount)
             expect(
               container.querySelector(".m-vehicle-map__route-shape")
             ).toBeInTheDocument()
@@ -823,14 +829,10 @@ describe("<MapPage />", () => {
 
           const selectedVehicle = shuttleFactory.build()
 
-          const shape = shapeFactory.build({
-            stops: stopFactory.buildList(8),
-          })
-
           setHtmlWidthHeightForLeafletMap()
           mockUseVehiclesForRouteMap({})
           mockUseVehicleForId([selectedVehicle])
-          mockUseTripShape({ [selectedVehicle.tripId!]: shape })
+          mockRoutePatternsForVehicles([]) // no route pattern for shuttle vehicle
 
           const { container } = render(
             <StateDispatchProvider
